refactor(api): move FakeLevel from yesman.epicfight.api.utils -> yesman.epicfight.client.world.util
